#118 BackgroundOverlayの改修
