+++ src/assets/sass/_variables.scss
@@ -2,6 +2,8 @@
 $color_timerBackground: #F3F2F9;
 
 // z-indexの指定
+$zIndex_buttonGroup: 120;
+$zIndex_bottomSheet: 110;
 $zIndex_backgroundOverlay: 100;
 $zIndex_loadingBar: 20;
 $zIndex_timerStartButton: 20;
