+++ src/components/organisms/DeleteButtonGroup.vue
@@ -1,0 +1,47 @@
+<template>
+  <div class="DeleteButtonGroup">
+    <BackgroundOverlay class="DeleteButtonGroup_BackgroundOverlay" />
+    <div class="DeleteButtonGroup_Inner">
+      <ApplyButton class="DeleteButtonGroup_Button">Delete</ApplyButton>
+      <CancelButton class="DeleteButtonGroup_Button">Cancel</CancelButton>
+    </div>
+  </div>
+</template>
+
+<script lang="ts">
+import { Component, Vue } from 'vue-property-decorator';
+import ApplyButton from '~/atoms/ApplyButton.vue';
+import BackgroundOverlay from '~/atoms/BackgroundOverlay.vue';
+import CancelButton from '~/atoms/CancelButton.vue';
+
+@Component({
+  components: {
+    ApplyButton,
+    BackgroundOverlay,
+    CancelButton,
+  },
+})
+export default class DeleteButtonGroup extends Vue {
+}
+</script>
+
+<style lang="scss" scoped>
+.DeleteButtonGroup {
+  &_BackgroundOverlay {
+    background: rgba(#000, 0.2);
+  }
+
+  &_Inner {
+    position: fixed;
+    bottom: 0;
+    z-index: $zIndex_buttonGroup;
+    width: 100%;
+    padding: 0 8px 8px 8px;
+    box-sizing: border-box;
+  }
+
+  &_Button + &_Button {
+    margin-top: 8px;
+  }
+}
+</style>
