+++ src/pages/Timer.vue
@@ -19,6 +19,8 @@
     <TimerEditor v-if="bottomSheet.isShown('timerEditor')" />
     <ProjectSelector v-if="false" />
     <TagsSelector v-if="false" />
+    <DeleteButtonGroup v-if="false" />
+    <DiscardButtonGroup v-if="false" />
   </div>
 </template>
 
@@ -29,6 +31,8 @@ import LoadingBar from '~/atoms/LoadingBar.vue';
 import TimerStartButton from '~/atoms/TimerStartButton.vue';
 import ActiveTimer from '~/molecules/ActiveTimer.vue';
 import BaseContent from '~/organisms/BaseContent.vue';
+import DeleteButtonGroup from '~/organisms/DeleteButtonGroup.vue';
+import DiscardButtonGroup from '~/organisms/DiscardButtonGroup.vue';
 import GlobalHeader from '~/organisms/GlobalHeader.vue';
 import GlobalNav from '~/organisms/GlobalNav.vue';
 import ProjectSelector from '~/organisms/ProjectSelector.vue';
@@ -46,6 +50,8 @@ const loadingTime: number = 3000;
     TimerStartButton,
     ActiveTimer,
     BaseContent,
+    DeleteButtonGroup,
+    DiscardButtonGroup,
     GlobalHeader,
     GlobalNav,
     ProjectSelector,
